requirements.txt
cycler==0.10.0
easyocr==1.4.1
imageio==2.9.0
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.3
opencv-python-headless==4.5.4.58
Pillow==8.2.0
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==6.0
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
tifffile==2021.10.12
torch==1.9.1
torchvision==0.10.1
typing-extensions==3.10.0.2
pip==21.1.2
wheel==0.36.2
certifi==2022.12.7
requests==2.28.1
idna==3.4
urllib3==1.26.13
setuptools==57.0.0
Pillow==9.3.0
PySocks==1.7.1
PyWavelets==1.4.1
PyYAML==6.0
Shapely==1.8.5.post1
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
easyocr==1.6.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
imageio==2.22.4
logger==1.4
loguru==0.6.0
multidict==6.0.3
networkx==
ninja==
numpy==1.23.5
opencv-python-headless==4.5.4.60
packaging==22.0
pafy==0.5.5
pip==21.1.2
pyTelegramBotAPI==4.8.0
pyclipper==1.3.0.post4
python-bidi==0.4.2
pytube==12.1.2
requests==2.28.1
rfc3986==1.5.0
scikit-image==0.19.3
scipy==1.9.3
setuptools==57.0.0
six==1.16.0
sniffio==1.3.0
tifffile==2022.10.10
torch==1.13.0
torchvision==0.14.0
typing-extensions==4.4.0
urllib3==1.26.13
wheel==0.36.2
win32-setctime==1.1.0
yandex-music==2.0.1
yarl==1.8.2
youtube-dl==2021.12.17	
youtube-search-python==1.6.6

